requirements.txt
streamlit
google-genai
pymatgen
python-dotenv
pillow
stmol 
py3Dmol
ipython-genutils
